Move demangle helper to gridtools::common namespace
